Move search route
