+++ x-pack/plugins/enterprise_search/server/routes/app_search/index.ts
@@ -16,7 +16,6 @@ import { registerEnginesRoutes } from './engines';
 import { registerOnboardingRoutes } from './onboarding';
 import { registerResultSettingsRoutes } from './result_settings';
 import { registerRoleMappingsRoutes } from './role_mappings';
-import { registerSampleResponseSearchRoute } from './sample_response_search';
 import { registerSearchSettingsRoutes } from './search_settings';
 import { registerSettingsRoutes } from './settings';
 
@@ -29,7 +28,6 @@ export const registerAppSearchRoutes = (dependencies: RouteDependencies) => {
   registerDocumentRoutes(dependencies);
   registerCurationsRoutes(dependencies);
   registerSearchSettingsRoutes(dependencies);
-  registerSampleResponseSearchRoute(dependencies);
   registerRoleMappingsRoutes(dependencies);
   registerResultSettingsRoutes(dependencies);
   registerApiLogsRoutes(dependencies);
