+++ x-pack/plugins/enterprise_search/server/routes/app_search/result_settings.ts
@@ -45,4 +45,22 @@ export function registerResultSettingsRoutes({
       path: '/as/engines/:engineName/result_settings',
     })
   );
+
+  router.post(
+    {
+      path: '/api/app_search/engines/{engineName}/sample_response_search',
+      validate: {
+        params: schema.object({
+          engineName: schema.string(),
+        }),
+        body: schema.object({
+          query: schema.string(),
+          result_fields: schema.recordOf(schema.string(), schema.object({}, { unknowns: 'allow' })),
+        }),
+      },
+    },
+    enterpriseSearchRequestHandler.createRequest({
+      path: '/as/engines/:engineName/sample_response_search',
+    })
+  );
 }
